+++ src/Firebase/Messaging/WebPushConfig.php
@@ -5,12 +5,12 @@ declare(strict_types=1);
 namespace Kreait\Firebase\Messaging;
 
 use JsonSerializable;
+use Kreait\Firebase\Exception\Messaging\InvalidArgument;
 
 /**
  * @see https://tools.ietf.org/html/rfc8030#section-5.3 Web Push Message Urgency
- *
  * @phpstan-type WebPushHeadersShape array{
- *     TTL?: positive-int,
+ *     TTL?: positive-int|numeric-string,
  *     Urgency?: self::URGENCY_*
  * }
  *
@@ -59,6 +59,13 @@ final class WebPushConfig implements JsonSerializable
     private const URGENCY_NORMAL = 'normal';
     private const URGENCY_HIGH = 'high';
 
+    private const VALID_URGENCIES = [
+        self::URGENCY_VERY_LOW,
+        self::URGENCY_LOW,
+        self::URGENCY_NORMAL,
+        self::URGENCY_HIGH,
+    ];
+
     /**
      * @var WebPushConfigShape
      */
@@ -82,7 +89,42 @@ final class WebPushConfig implements JsonSerializable
      */
     public static function fromArray(array $config): self
     {
+        if (array_key_exists('headers', $config) && is_array($config['headers'])) {
+            $config['headers'] = self::ensureValidHeaders($config['headers']);
+        }
+
         return new self($config);
+    }
+
+    /**
+     * @param WebPushHeadersShape $headers
+     *
+     * @return WebPushHeadersShape
+     */
+    private static function ensureValidHeaders(array $headers): array
+    {
+        if (array_key_exists('TTL', $headers) && is_int($headers['TTL'])) {
+            $headers['TTL'] = (string) $headers['TTL'];
+        }
+
+        if (
+            array_key_exists('TTL', $headers)
+            && is_string($headers['TTL'])
+            && preg_match('/^[\-0]/', $headers['TTL']) === 1
+        ) {
+            throw new InvalidArgument('The TTL in the WebPushConfig must must be a positive int');
+        }
+
+        if (array_key_exists('Urgency', $headers)) {
+            if (!in_array($headers['Urgency'], self::VALID_URGENCIES, true)) {
+                throw new InvalidArgument(sprintf(
+                    'The Urgency in the WebPushConfig header must must be one of %s',
+                    implode(',', self::VALID_URGENCIES)
+                ));
+            }
+        }
+
+        return $headers;
     }
 
     public function withHighUrgency(): self